+++ packages/mattermost-plugin/components/InviteToMeetingModal/InviteToMeetingModal.tsx
@@ -18,11 +18,18 @@ const InviteToMeetingModal = () => {
   const channel = useCurrentChannel()
   const data = useLazyLoadQuery<InviteToMeetingModalQuery>(
     graphql`
-      query InviteToMeetingModalQuery($channel: ID!) {
-        linkedTeamIds(channel: $channel)
+      query InviteToMeetingModalQuery {
         viewer {
           teams {
             id
+            viewerTeamMember {
+              id
+              integrations {
+                mattermost {
+                  linkedChannels
+                }
+              }
+            }
             activeMeetings {
               ...useInviteToMeeting_meeting
               id
@@ -32,17 +39,21 @@ const InviteToMeetingModal = () => {
         }
       }
     `,
-    {
-      channel: channel?.id ?? ''
-    }
+    {}
   )
 
-  const {viewer, linkedTeamIds} = data
+  const linkedTeams = useMemo(() => {
+    const {viewer} = data
+    return viewer.teams.filter(
+      (team) =>
+        channel &&
+        team.viewerTeamMember?.integrations.mattermost.linkedChannels.includes(channel.id)
+    )
+  }, [data, channel])
+
+  const {viewer} = data
   const {teams} = viewer
-  const linkedTeams = useMemo(
-    () => viewer.teams.filter((team) => !linkedTeamIds || linkedTeamIds.includes(team.id)),
-    [viewer, linkedTeamIds]
-  )
+
   const activeMeetings = useMemo(
     () => linkedTeams.flatMap((team) => team.activeMeetings),
     [linkedTeams]